Changing the project structure
