+++ .settings/org.eclipse.jpt.core.prefs
@@ -1,3 +1,4 @@
 eclipse.preferences.version=1
+entitygen.DEFAULT_PACKAGE=model
 org.eclipse.jpt.core.platform=generic2_2
-org.eclipse.jpt.jpa.core.discoverAnnotatedClasses=false
+org.eclipse.jpt.jpa.core.discoverAnnotatedClasses=true
