+++ src/main/java/com/greenbus/MusicAppApplication.java
@@ -1,0 +1,21 @@
+package com.greenbus;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
+
+//import config.WebSecurityConfig;
+
+
+
+@SpringBootApplication
+@EnableAutoConfiguration(exclude = {org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration.class})
+@ComponentScan(value = "com.greenbus")// This annotation enables component scanning in its package and sub-packages
+public class MusicAppApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MusicAppApplication.class, args);
+    }
+}
