+++ src/main/java/com/greenbus/config/WebSecurityConfig.java
@@ -1,0 +1,50 @@
+package com.greenbus.config;
+
+import java.util.Arrays;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+
+
+@Configuration
+@EnableWebSecurity
+public class WebSecurityConfig{
+	
+	@Bean
+	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+		
+		http.authorizeRequests().anyRequest().permitAll();
+		http.csrf().disable();
+		
+		return http.build();
+	}
+	@Bean
+	public InMemoryUserDetailsManager userDetailsService() {
+	    UserDetails user = User.withUsername("cst8277")
+	      .password("8277")
+	      .roles("USER")
+	      .build();
+	    return new InMemoryUserDetailsManager(user);
+	}
+	
+	
+    @Bean
+    public BCryptPasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+    
+ 
+}
