--- src/main/java/config/WebSecurityConfig.java
@@ -1,14 +1,0 @@
-package config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class WebSecurityConfig {
-
-    @Bean
-    public BCryptPasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
