+++ src/main/resources/application.properties
@@ -1,12 +1,28 @@
 # DataSource settings
 spring.datasource.url=jdbc:mysql://localhost:3306/GreenBusMusicDB?useSSL=false&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=1234
+spring.datasource.username=cst8277
+spring.datasource.password=8277
+# Define the driver class (if necessary)
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # JPA / Hibernate settings
 spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
+#ddl-auto values: create, create-drop, validate, and update
+spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
+management.security.enabled=false
+# spring.h2.console.enabled=true
+spring.security.basic.enable: false
+spring.security.user.name=cst8277 
+spring.security.user.password=8277
 
-# Define the driver class (if necessary)
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+log4j.logger.org.hibernate.SQL=DEBUG
+log4j.logger.org.hibernate.type=TRACE
+logging.level.root=info
+logging.level.druid.sql.statement=debug
+logging.level.com.example=debug
+
+music.local.song=D:/Term4/CST8319 Software Development Project/upload/song
+music.local.cover=D:/Term4/CST8319 Software Development Project/upload/cover
